working on legal moves for king
